Sentinel mixin creation and folder restructure
