+++ msticpy/data/azure/__init__.py
@@ -1,0 +1,9 @@
+# -------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for
+# license information.
+# --------------------------------------------------------------------------
+"""Data provider sub-package."""
+
+from .azure_data import AzureData
+from .sentinel_core import MicrosoftSentinel
